+++ src/components/LoadMore.ts
@@ -1,87 +1,5 @@
-/* eslint-disable max-lines-per-function */
-import { movieDataStateStore } from "../model";
-import {
-  handleGetPopularMovieData,
-  handleGetSearchMovieData,
-} from "../service/handleSkeletonAndAPI";
 import { ListType } from "../type/movie";
-
-import ItemCardList from "./ItemCardList";
-
-const getSearchInputValue = () => {
-  const $searchInput = document.querySelector("#search-input");
-
-  if (!($searchInput instanceof HTMLInputElement)) {
-    return undefined;
-  }
-
-  return $searchInput.value;
-};
-
-const getSearchMovieData = async () => {
-  const title = getSearchInputValue();
-
-  if (!title) return;
-
-  await handleGetSearchMovieData(title, false);
-};
-
-const addItemCardList = () => {
-  if (
-    movieDataStateStore.fetchedMovieData.movieList !== undefined &&
-    movieDataStateStore.fetchedMovieData.movieList.length !== 0
-  ) {
-    ItemCardList(movieDataStateStore.fetchedMovieData.movieList);
-  }
-
-  if (movieDataStateStore.fetchedMovieData.isEndPage) {
-    const $loader = document.querySelector(".loader");
-    $loader?.remove();
-  }
-};
-
-const fetchMoreData = async (listType: ListType) => {
-  if (listType === "popular") {
-    await handleGetPopularMovieData();
-  } else {
-    await getSearchMovieData();
-  }
-  addItemCardList();
-};
-
-const handleIntersection = (
-  entries: IntersectionObserverEntry[],
-  listType: ListType,
-) => {
-  entries.forEach((entry) => {
-    if (entry.isIntersecting) {
-      fetchMoreData(listType);
-    }
-  });
-};
-
-const createIntersectionObserver = (listType: ListType) => {
-  const observer = new IntersectionObserver(
-    (entries) => handleIntersection(entries, listType),
-    {
-      rootMargin: "0px 0px -50px 0px",
-    },
-  );
-
-  return observer;
-};
-
-let previousObserver: IntersectionObserver | null = null;
-const addObserver = ($loadMore: HTMLElement, listType: ListType) => {
-  if (previousObserver) {
-    previousObserver.disconnect();
-  }
-
-  const observer = createIntersectionObserver(listType);
-  observer.observe($loadMore);
-
-  previousObserver = observer;
-};
+import addObserver from "../utils/IntersectionObserver";
 
 const LoadMore = (listType: ListType, isEndPage: boolean) => {
   const $itemView = document.querySelector(".item-view");
